% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blank_data.R
\name{combineBlanks}
\alias{combineBlanks}
\title{combineBlanks}
\usage{
combineBlanks(raw, norm)
}
\arguments{
\item{raw}{a raw blank datatable from getRaw}

\item{norm}{A normalized blank datatable from getNorm}
}
\value{
A data table of combined blank data
}
\description{
combineBlanks
}
