% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blank_data.R
\name{getBlankData}
\alias{getBlankData}
\title{getBlankData}
\usage{
getBlankData(from = "2021-01-01")
}
\arguments{
\item{from}{A date object}
}
\value{
A data table of combined blank data
}
\description{
getBlankData
}
