% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blank_data.R
\name{getNormBlankWheels}
\alias{getNormBlankWheels}
\title{getNormBlankWheels}
\usage{
getNormBlankWheels(wheels, recs = c(83028, 53804, 2138, 140548, 36168,
  55101, 1081, 39246, 32490, 32491, 32492, 36947, 148820, 159579))
}
\arguments{
\item{wheels}{A vector of wheels}

\item{recs}{A vector of rec_nums}
}
\value{
A data table of normalized blank data
}
\description{
getNormBlankWheels
}
