% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blank_data.R
\name{getRaw}
\alias{getRaw}
\title{getRaw}
\usage{
getRaw(
  from = as.Date("2021-01-01"),
  recs = c(83028, 53804, 2138, 140548, 36168, 55101, 1081, 39246, 32490, 32491, 32492,
    36947, 64052, 148820, 159579, 178222)
)
}
\arguments{
\item{from}{A date object}

\item{recs}{A vector of rec_nums}
}
\value{
A data table of raw blank data
}
\description{
getRaw
}
